requirements.txt
# click + console
click==7.1.2
colorama~=0.4.4  # color logging windows
inquirer~=2.7.0
tabulate~=0.8.7

# libraries
pydantic~=1.7.3
pyyaml~=5.3.1
tinydb~=3.15.2

requests~=2.25.1
pyjwt[crypto]~=2.0.1

gql~=0.4.0
semantic-version~=2.8.4
kubernetes>=11.0.0
retrying~=1.3.3
